Raise exceptions on version generation in debug
If you do not have PIL installed correctly and `version_generator` throws an exception, it is disregarded here, even in when `TEMPLATE_DEBUG` is `True`.  This makes it hard to determine why you can't select a different size version in the admin.  What is passed back to the image selection window is a blank string.  Re-raising the exception here is helpful for debugging.