Update custom_field.html
It's possible someone could manually change the path before trying to generate the thumbnail that this check protects.

Simple change to double check that "value.exists" before trying to generate a thumbnail.

I was seeing internal server errors upon save if I chose a file, then manually updated the path to it, or otherwise just put a nonsense path in the input field.  This patch fixed it.